requirements.txt
torch>=1.9.0
transformers>=4.20.0
Pillow>=8.0.0
pandas>=1.3.0
